Put Add and Remove files in different modules
